[WIP][Object Storage] Update: test
